% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_of_quantiles.R
\name{quantile_of_quantiles}
\alias{quantile_of_quantiles}
\title{quantil_of_quantiles}
\usage{
quantile_of_quantiles(df, inputs = NULL)
}
\arguments{
\item{df}{data frame}

\item{inputs}{character vector with the names of the columns to be converted
to a quantile of quantiles}
}
\value{
numeric vector of quantiles
}
\description{
Calculate quantiles for each input, sum the input quantiles, and calculate
quantile of the sum.
}
\details{
Note: NA values in any of the input vectors will result in a NA value for the
quantile of quantiles.
}
